로그인 기능 구현

+++ src/main/java/com/sbs/basic1/boundedContext/home/controller/HomeController.java
@@ -238,19 +238,19 @@ public class HomeController {
 
     @GetMapping("/home/cookie/increase")
     @ResponseBody
-    public int showCookieIncrease(HttpServletRequest req, HttpServletResponse resp) {
-        int countInCookie = 0;
+    public long showCookieIncrease(HttpServletRequest req, HttpServletResponse resp) {
+        long countInCookie = 0;
 
         if (req.getCookies() != null) {
             countInCookie = Arrays.stream(req.getCookies())
                     .filter(cookie -> cookie.getName().equals("count"))
                     .map(cookie -> cookie.getValue())
-                    .mapToInt(Integer::parseInt)
+                    .mapToLong(Long::parseLong)
                     .findFirst()
                     .orElse(0);
         }
 
-        int newCountInCookie = countInCookie + 1;
+        long newCountInCookie = countInCookie + 1;
 
         resp.addCookie(new Cookie("count", newCountInCookie + ""));
